feat: prototype stronger feedback typings
