+++ src/actions/index.ts
@@ -1,30 +1,30 @@
 import { type Atem } from 'atem-connection'
-import type { AtemConfig } from './config.js'
-import type { ModelSpec } from './models/index.js'
-import { type InstanceBaseExt } from './util.js'
-import { AtemCommandBatching } from './batching.js'
-import { AtemTransitions } from './transitions.js'
+import type { AtemConfig } from '../config.js'
+import type { ModelSpec } from '../models/index.js'
+import { type InstanceBaseExt } from '../util.js'
+import { AtemCommandBatching } from '../batching.js'
+import { AtemTransitions } from '../transitions.js'
 import type { CompanionActionDefinition } from '@companion-module/base'
-import { createProgramPreviewActions } from './actions/mixeffect/programPreview.js'
-import { convertMyActionDefinitions } from './actions/wrapper.js'
-import { createTransitionActions } from './actions/mixeffect/transition.js'
-import { createDisplayClockActions } from './actions/displayClock.js'
-import { createMacroActions } from './actions/macro.js'
-import { createStreamingActions } from './actions/streaming.js'
-import { createRecordingActions } from './actions/recording.js'
-import { createDownstreamKeyerActions } from './actions/dsk.js'
-import { createAuxOutputActions } from './actions/aux.js'
-import { createMultiviewerActions } from './actions/multiviewer.js'
-import { createMediaPlayerActions } from './actions/mediaPlayer.js'
-import { createSettingsActions } from './actions/settings.js'
-import { createSuperSourceActions } from './actions/superSource.js'
-import { createUpstreamKeyerCommonActions } from './actions/mixeffect/upstreamKeyerCommon.js'
-import { createFadeToBlackActions } from './actions/mixeffect/fadeToBlack.js'
-import { createUpstreamKeyerDVEActions } from './actions/mixeffect/upstreamKeyerDVE.js'
-import { createClassicAudioActions } from './actions/classicAudio.js'
-import { createFairlightAudioActions } from './actions/fairlightAudio.js'
-import type { MyActionDefinition } from './actions/types.js'
-import type { StateWrapper } from './state.js'
+import { createProgramPreviewActions } from './mixeffect/programPreview.js'
+import { convertMyActionDefinitions } from './wrapper.js'
+import { createTransitionActions } from './mixeffect/transition.js'
+import { createDisplayClockActions } from './displayClock.js'
+import { createMacroActions } from './macro.js'
+import { createStreamingActions } from './streaming.js'
+import { createRecordingActions } from './recording.js'
+import { createDownstreamKeyerActions } from './dsk.js'
+import { createAuxOutputActions } from './aux.js'
+import { createMultiviewerActions } from './multiviewer.js'
+import { createMediaPlayerActions } from './mediaPlayer.js'
+import { createSettingsActions } from './settings.js'
+import { createSuperSourceActions } from './superSource.js'
+import { createUpstreamKeyerCommonActions } from './mixeffect/upstreamKeyerCommon.js'
+import { createFadeToBlackActions } from './mixeffect/fadeToBlack.js'
+import { createUpstreamKeyerDVEActions } from './mixeffect/upstreamKeyerDVE.js'
+import { createClassicAudioActions } from './classicAudio.js'
+import { createFairlightAudioActions } from './fairlightAudio.js'
+import type { MyActionDefinition } from './types.js'
+import type { StateWrapper } from '../state.js'
 
 export enum ActionId {
 	Program = 'program',
