+++ src/actions/mixeffect/programPreview.ts
@@ -2,7 +2,7 @@ import type { Atem } from 'atem-connection'
 import { getMixEffect } from 'atem-connection/dist/state/util.js'
 import { AtemMEPicker, AtemMESourcePicker, FadeDurationChoice } from '../../input.js'
 import type { ModelSpec } from '../../models/index.js'
-import { ActionId } from '../../actions.js'
+import { ActionId } from '../index.js'
 import type { MyActionDefinitions } from '../types.js'
 import type { StateWrapper } from '../../state.js'
 import type { AtemTransitions } from '../../transitions.js'
