+++ src/actions/mixeffect/upstreamKeyerDVE.ts
@@ -1,7 +1,7 @@
 import { Enums, type Atem } from 'atem-connection'
 import { AtemMEPicker, AtemUSKDVEPropertiesPickers, AtemUSKPicker } from '../../input.js'
 import type { ModelSpec } from '../../models/index.js'
-import { ActionId } from '../../actions.js'
+import { ActionId } from '../index.js'
 import type { MyActionDefinitions } from '../types.js'
 import { CHOICES_FLYDIRECTIONS, CHOICES_KEYFRAMES } from '../../choices.js'
 import { getUSK, type StateWrapper } from '../../state.js'
