+++ src/actions/types.ts
@@ -1,33 +1,6 @@
 import type { CompanionActionContext, SomeCompanionActionInputField } from '@companion-module/base'
-import type { ConditionalKeys } from 'type-fest'
+import type { MyOptionsHelper, MyOptionsObject } from '../common'
 
-// type InputFieldNumber = CompanionInputFieldColor | CompanionInputFieldNumber
-// type InputFieldString = CompanionInputFieldTextInput
-
-// type PrimitiveToActionInputField<T> = T extends boolean
-// 	? CompanionInputFieldCheckbox
-// 	: T extends number | string
-// 	? CompanionInputFieldDropdown | (T extends string ? InputFieldString : T extends number ? InputFieldNumber : never)
-// 	: never
-
-export type MyActionOptionsObject<TOptions> = {
-	[K in keyof TOptions]: undefined extends TOptions[K]
-		? SomeCompanionActionInputField | undefined
-		: SomeCompanionActionInputField
-	// [K in keyof TOptions]: PrimitiveToActionInputField<TOptions[K]>
-	// [k: string]: CompanionInputFieldStaticText
-}
-
-export interface MyOptionsHelper<TOptions> {
-	getJson(): TOptions
-	getRaw<Key extends keyof TOptions>(fieldName: Key): TOptions[Key] | undefined
-	getPlainString<Key extends ConditionalKeys<TOptions, string>>(fieldName: Key): TOptions[Key]
-	getPlainNumber<Key extends ConditionalKeys<TOptions, number>>(fieldName: Key): TOptions[Key]
-	getPlainBoolean<Key extends ConditionalKeys<TOptions, boolean>>(fieldName: Key): boolean
-
-	getParsedString<Key extends ConditionalKeys<TOptions, string | undefined>>(fieldName: Key): Promise<string>
-	getParsedNumber<Key extends ConditionalKeys<TOptions, string | undefined>>(fieldName: Key): Promise<number>
-}
 /**
  * Basic information about an instance of an action
  */
@@ -55,7 +28,7 @@ export interface MyActionDefinition<TOptions> {
 	/** Additional description of the action */
 	description?: string
 	/** The input fields for the action */
-	options: MyActionOptionsObject<TOptions>
+	options: MyOptionsObject<TOptions, SomeCompanionActionInputField>
 
 	/** Called to execute the action */
 	callback: (action: MyActionEvent2<TOptions>, context: CompanionActionContext) => Promise<void> | void
