+++ src/feedback/multiviewer.ts
@@ -1,48 +1,54 @@
-import { type Atem } from 'atem-connection'
+import { AtemMultiviewSourcePicker, AtemMultiviewWindowPicker, AtemMultiviewerPicker } from '../input.js'
 import type { ModelSpec } from '../models/index.js'
-import { ActionId } from './index.js'
-import type { MyActionDefinitions } from './types.js'
+import type { MyFeedbackDefinitions } from './types.js'
+import { FeedbackId } from './index.js'
+import { combineRgb } from '@companion-module/base'
 import { getMultiviewerWindow, type StateWrapper } from '../state.js'
-import { AtemMultiviewerPicker, AtemMultiviewWindowPicker, AtemMultiviewSourcePicker } from '../input.js'
 
-export interface AtemMultiviewerActions {
-	[ActionId.MultiviewerWindowSource]: {
+export interface AtemMultiviewerFeedbacks {
+	[FeedbackId.MVSource]: {
 		multiViewerId: number
 		windowIndex: number
 		source: number
 	}
-	[ActionId.MultiviewerWindowSourceVariables]: {
+	[FeedbackId.MVSourceVariables]: {
 		multiViewerId: string
 		windowIndex: string
 		source: string
 	}
 }
 
-export function createMultiviewerActions(
-	atem: Atem | undefined,
+export function createMultiviewerFeedbacks(
 	model: ModelSpec,
 	state: StateWrapper,
-): MyActionDefinitions<AtemMultiviewerActions> {
+): MyFeedbackDefinitions<AtemMultiviewerFeedbacks> {
 	if (!model.MVs) {
 		return {
-			[ActionId.MultiviewerWindowSource]: undefined,
-			[ActionId.MultiviewerWindowSourceVariables]: undefined,
+			[FeedbackId.MVSource]: undefined,
+			[FeedbackId.MVSourceVariables]: undefined,
 		}
 	}
 	return {
-		[ActionId.MultiviewerWindowSource]: {
-			name: 'Multiviewer: Change window source',
+		[FeedbackId.MVSource]: {
+			type: 'boolean',
+			name: 'Multiviewer: Window source',
+			description: 'If the specified MV window is set to the specified source, change style of the bank',
 			options: {
 				multiViewerId: AtemMultiviewerPicker(model),
 				windowIndex: AtemMultiviewWindowPicker(model),
 				source: AtemMultiviewSourcePicker(model, state.state),
 			},
-			callback: async ({ options }) => {
-				await atem?.setMultiViewerWindowSource(
-					options.getPlainNumber('source'),
+			defaultStyle: {
+				color: combineRgb(0, 0, 0),
+				bgcolor: combineRgb(255, 255, 0),
+			},
+			callback: ({ options }): boolean => {
+				const window = getMultiviewerWindow(
+					state.state,
 					options.getPlainNumber('multiViewerId'),
 					options.getPlainNumber('windowIndex'),
 				)
+				return window?.source === options.getPlainNumber('source')
 			},
 			learn: ({ options }) => {
 				const window = getMultiviewerWindow(
@@ -61,8 +67,10 @@ export function createMultiviewerActions(
 				}
 			},
 		},
-		[ActionId.MultiviewerWindowSourceVariables]: {
-			name: 'Multiviewer: Change window source from variables',
+		[FeedbackId.MVSourceVariables]: {
+			type: 'boolean',
+			name: 'Multiviewer: Window source from variables',
+			description: 'If the specified MV window is set to the specified source, change style of the bank',
 			options: {
 				multiViewerId: {
 					type: 'textinput',
@@ -86,14 +94,17 @@ export function createMultiviewerActions(
 					useVariables: true,
 				},
 			},
+			defaultStyle: {
+				color: combineRgb(0, 0, 0),
+				bgcolor: combineRgb(255, 255, 0),
+			},
 			callback: async ({ options }) => {
 				const multiViewerId = (await options.getParsedNumber('multiViewerId')) - 1
 				const windowIndex = (await options.getParsedNumber('windowIndex')) - 1
 				const source = await options.getParsedNumber('source')
 
-				if (isNaN(multiViewerId) || isNaN(windowIndex) || isNaN(source)) return
-
-				await atem?.setMultiViewerWindowSource(source, multiViewerId, windowIndex)
+				const window = getMultiviewerWindow(state.state, multiViewerId, windowIndex)
+				return window?.source === source
 			},
 			learn: async ({ options }) => {
 				const multiViewerId = (await options.getParsedNumber('multiViewerId')) - 1
